backend/requirements.txt
asgiref==3.5.2
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.12
django-comment-migrate==0.1.7
django-cors-headers==3.10.1
django-filter==22.1
django-ranged-response==0.2.0
django-redis==5.2.0
django-restql==0.15.3
django-simple-captcha==0.5.17
django-tenants==3.4.8
django-timezone-field==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
packaging==23.0
drf-yasg==1.21.5
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.0.1
mysqlclient==2.1.1
Pillow==9.4.0
PyJWT==2.6.0
pyparsing==2.4.7
pyPEG2==2.15.2
pypinyin==0.48.0
pytz==2021.1
requests==2.28.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
six==1.16.0
smmap==4.0.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
tzlocal==2.1
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.15
user-agents==2.2.0
whitenoise==5.3.0
openpyxl==3.1.2
channels==3.0.5
channels-redis==3.4.1
uvicorn==0.21.1
gunicorn==20.1.0
gevent==22.10.2
websockets==10.4
# project new packages
argilla
# elasticsearch